agent/requirements.txt
langgraph>=0.2.28
langchain-groq>=0.1.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
